+++ src/Order/Changelog.md
@@ -7,7 +7,8 @@
 
 - Dropped PHP 8.2 Support
 - Changed the minimum Laravel 10 version to v10.48
-- Changed the `Order` and `OrderItem` interfaces to extend the new `Sale` and `SaleItem` interfaces from the contracts module (the resulting interface signature is unchanged)
+- BC: Added the `getBuyable()` method to the `OrderItem` interface
+- Changed the `Order` and `OrderItem` interfaces to extend the new `Sale` and `SaleItem` interfaces from the contracts module
 - Added Laravel 12 Support
 - Fixed the missing float cast of the order item's price field
 - PHP 8.4 deprecation notice fixes